nb/pyproject.toml
[project]
name = "nb"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Quinn Dougherty", email = "quinnd@riseup.net" }
]
dependencies = [
    "datasets>=2.21.0",
    "anthropic>=0.34.2",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "jupyter>=1.1.1",
    "black>=24.8.0",
    "pytest>=8.3.2",
    "hypothesis>=6.112.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/nb"]
